requirements.txt
elasticsearch>=6.3.1
kafka-python>=1.4.4
myconf>=1.1.0
